Check login #37

+++ src/main/java/com/example/diary/controller/DiaryDeleteController.java
@@ -18,6 +18,7 @@ public class DiaryDeleteController {
 
     @PostMapping("/diaryDeleteCheck")
     public String postDiaryDeleteCheck(@ModelAttribute StudentDiaryForm studentDiaryForm, Model model) {
+        if (!diaryService.checkLogin()) return "sessionError";
         diaryService.addContentsAndTitle(model, "diaryDeleteCheck", "日誌削除確認");
 
         return "student/main";
@@ -25,6 +26,7 @@ public class DiaryDeleteController {
 
     @PostMapping("/diaryDeleteComplete")
     public String postDiaryDeleteComplete(@ModelAttribute StudentDiaryForm studentDiaryForm, HttpSession session, Model model) {
+        if (!diaryService.checkLogin()) return "sessionError";
         diaryService.addContentsAndTitle(model, "diaryDeleteComplete", "日誌削除完了");
 
         int row = diaryService.deleteDiary(diaryService.setDiaryClass(studentDiaryForm, session));
